refactor(server): split MCP provider internals
Closes #191
