+++ apps/server/src/tool/mcp/provider-metadata.ts
@@ -1,0 +1,37 @@
+import type { RuntimeResource, Tool } from "@openomni/protocol";
+
+function uniqueLabels(labels: readonly string[]): string[] {
+  return [...new Set(labels)];
+}
+
+/** @internal Package-local helper for McpToolProvider only. */
+export function mcpToolMetadata(
+  serverName: string,
+  spec: Tool.Spec,
+): { readonly labels: string[]; readonly descriptor: RuntimeResource.Descriptor } {
+  const labels = uniqueLabels([
+    ...(spec.labels ?? []),
+    `tool:${spec.name}`,
+    "source.mcp",
+    `mcp.${serverName}`,
+  ]);
+
+  return {
+    labels,
+    descriptor: {
+      id: `tool:mcp:${serverName}:${spec.name}`,
+      kind: "tool",
+      source: { type: "mcp", serverId: serverName, remoteName: spec.name },
+      labels,
+      capabilities: [],
+      effects: [],
+    },
+  };
+}
+
+/** @internal Package-local helper for McpToolProvider only. */
+export function createResultSummary(output: unknown): string {
+  const outputStr = typeof output === "string" ? output : JSON.stringify(output);
+  const length = outputStr.length;
+  return `success:text:${length}b`;
+}
