+++ apps/server/src/tool/mcp/provider.ts
@@ -1,8 +1,8 @@
 import { McpClient } from "@openomni/agent";
-import type { McpServerConfig, RuntimeResource, Tool } from "@openomni/protocol";
+import type { McpServerConfig, Tool } from "@openomni/protocol";
 import { Mcp, PolicyDecision, PolicyEvent, ToolExecution } from "@openomni/protocol";
 import { Operational } from "@openomni/protocol";
-import { Bus, Storage } from "@openomni/session";
+import { Bus } from "@openomni/session";
 import type {
   NativeTool,
   ToolCategory,
@@ -10,63 +10,29 @@ import type {
   ToolProvider,
 } from "@openomni/openomni";
 import { McpPrefixGuardMiddleware } from "./mcp-prefix-guard";
+import {
+  MCP_TOOL_ACTION,
+  buildActor,
+  publishLifecycleApproved,
+  publishLifecycleBlocked,
+  publishLifecycleRequested,
+  readSessionId,
+  resolveLifecycleAudit,
+  summarizeServerConfig,
+} from "./provider-audit";
+import { createResultSummary, mcpToolMetadata } from "./provider-metadata";
+import type {
+  McpClientLike,
+  McpLifecycleAuditContext,
+  McpToolProviderOptions,
+} from "./provider-types";
 
-const MCP_TOOL_ACTION = "mcp.tool.call";
+export type { McpLifecycleAuditContext, McpToolProviderOptions } from "./provider-types";
+
 function createAbortError(): Error {
   const error = new Error("MCP tool execution aborted");
   error.name = "AbortError";
   return error;
-}
-
-interface McpClientLike {
-  connect(): Promise<void>;
-  disconnect(): Promise<void>;
-  listTools(): Promise<Tool.Spec[]>;
-  callTool(
-    toolName: string,
-    input: Record<string, unknown>,
-    callId?: string,
-    context?: ToolExecutionContext,
-  ): Promise<Tool.Result>;
-}
-
-function uniqueLabels(labels: readonly string[]): string[] {
-  return [...new Set(labels)];
-}
-
-function mcpToolMetadata(
-  serverName: string,
-  spec: Tool.Spec,
-): { readonly labels: string[]; readonly descriptor: RuntimeResource.Descriptor } {
-  const labels = uniqueLabels([
-    ...(spec.labels ?? []),
-    `tool:${spec.name}`,
-    "source.mcp",
-    `mcp.${serverName}`,
-  ]);
-
-  return {
-    labels,
-    descriptor: {
-      id: `tool:mcp:${serverName}:${spec.name}`,
-      kind: "tool",
-      source: { type: "mcp", serverId: serverName, remoteName: spec.name },
-      labels,
-      capabilities: [],
-      effects: [],
-    },
-  };
-}
-
-export interface McpToolProviderOptions {
-  readonly createClient?: (config: McpServerConfig) => McpClientLike;
-}
-
-export interface McpLifecycleAuditContext {
-  readonly audit?: {
-    readonly sessionId?: string;
-  };
-  readonly actor?: Record<string, unknown>;
 }
 
 export class McpToolProvider implements ToolProvider {
@@ -84,10 +50,9 @@ export class McpToolProvider implements ToolProvider {
     const actionId = crypto.randomUUID();
 
     if (audit) {
-      Bus.publish(PolicyEvent.ActionRequested, {
-        ...lifecycleBase(audit),
+      publishLifecycleRequested({
+        audit,
         actionId,
-        actor: buildActor(audit.sessionId, audit.actor),
         action: "mcp.server.connect",
         resource: config.name,
         context: summarizeServerConfig(config),
@@ -102,25 +67,21 @@ export class McpToolProvider implements ToolProvider {
       this.cachedTools = null;
 
       if (audit) {
-        Bus.publish(PolicyEvent.ActionApproved, {
-          ...lifecycleBase(audit),
+        publishLifecycleApproved({
+          audit,
           actionId,
-          actor: buildActor(audit.sessionId, audit.actor),
           action: "mcp.server.connect",
           resource: config.name,
-          verdict: "allow" as const,
           reason: "MCP server connected",
         });
       }
     } catch (err) {
       if (audit) {
-        Bus.publish(PolicyEvent.ActionBlocked, {
-          ...lifecycleBase(audit),
+        publishLifecycleBlocked({
+          audit,
           actionId,
-          actor: buildActor(audit.sessionId, audit.actor),
           action: "mcp.server.connect",
           resource: config.name,
-          verdict: "deny" as const,
           reason: err instanceof Error ? err.message : String(err),
         });
       }
@@ -144,10 +105,9 @@ export class McpToolProvider implements ToolProvider {
     const actionId = crypto.randomUUID();
 
     if (audit) {
-      Bus.publish(PolicyEvent.ActionRequested, {
-        ...lifecycleBase(audit),
+      publishLifecycleRequested({
+        audit,
         actionId,
-        actor: buildActor(audit.sessionId, audit.actor),
         action: "mcp.server.disconnect",
         resource: name,
         context: { serverName: name },
@@ -160,13 +120,11 @@ export class McpToolProvider implements ToolProvider {
     this.cachedTools = null;
 
     if (audit) {
-      Bus.publish(PolicyEvent.ActionApproved, {
-        ...lifecycleBase(audit),
+      publishLifecycleApproved({
+        audit,
         actionId,
-        actor: buildActor(audit.sessionId, audit.actor),
         action: "mcp.server.disconnect",
         resource: name,
-        verdict: "allow" as const,
         reason: "MCP server disconnected",
       });
     }
@@ -178,10 +136,9 @@ export class McpToolProvider implements ToolProvider {
     const actionId = crypto.randomUUID();
 
     if (audit) {
-      Bus.publish(PolicyEvent.ActionRequested, {
-        ...lifecycleBase(audit),
+      publishLifecycleRequested({
+        audit,
         actionId,
-        actor: buildActor(audit.sessionId, audit.actor),
         action: "mcp.server.disconnect_all",
         resource: "mcp.servers",
         context: { serverNames },
@@ -197,13 +154,11 @@ export class McpToolProvider implements ToolProvider {
     this.cachedTools = null;
 
     if (audit) {
-      Bus.publish(PolicyEvent.ActionApproved, {
-        ...lifecycleBase(audit),
+      publishLifecycleApproved({
+        audit,
         actionId,
-        actor: buildActor(audit.sessionId, audit.actor),
         action: "mcp.server.disconnect_all",
         resource: "mcp.servers",
-        verdict: "allow" as const,
         reason: "MCP servers disconnected",
       });
     }
@@ -336,83 +291,4 @@ export class McpToolProvider implements ToolProvider {
   get serverCount(): number {
     return this.connected.size;
   }
-}
-
-type ResolvedLifecycleAudit = {
-  readonly sessionId: string;
-  readonly actor?: Record<string, unknown>;
-};
-
-function resolveLifecycleAudit(
-  context: McpLifecycleAuditContext | undefined,
-): ResolvedLifecycleAudit | undefined {
-  const sessionId = context?.audit?.sessionId;
-  if (typeof sessionId !== "string" || sessionId.length === 0) {
-    const fallbackSession = latestSession();
-    if (!fallbackSession) return undefined;
-    return {
-      sessionId: fallbackSession.id,
-      ...(context?.actor !== undefined && { actor: context.actor }),
-    };
-  }
-  return {
-    sessionId,
-    ...(context?.actor !== undefined && { actor: context.actor }),
-  };
-}
-
-function latestSession(): { readonly id: string } | undefined {
-  return Storage.get()
-    .session.list()
-    .sort((left, right) => right.time.updated - left.time.updated)[0];
-}
-
-function summarizeServerConfig(config: McpServerConfig): Record<string, unknown> {
-  const summary: Record<string, unknown> = {
-    serverName: config.name,
-    transport: config.transport,
-    ...(config.timeout !== undefined && { timeout: config.timeout }),
-    ...(config.retries !== undefined && { retries: config.retries }),
-    ...(config.headers !== undefined && { headerNames: Object.keys(config.headers).sort() }),
-  };
-
-  if (config.transport === "stdio") {
-    return {
-      ...summary,
-      command: config.command,
-      ...(config.args !== undefined && { argsCount: config.args.length }),
-    };
-  }
-
-  return { ...summary, url: config.url };
-}
-
-function lifecycleBase(audit: ResolvedLifecycleAudit) {
-  return {
-    traceId: crypto.randomUUID(),
-    sessionId: audit.sessionId,
-    time: Date.now(),
-  };
-}
-
-function readSessionId(call: Tool.Call): string | undefined {
-  const sessionId = call.input.sessionId;
-  return typeof sessionId === "string" && sessionId.length > 0 ? sessionId : undefined;
-}
-
-function buildActor(
-  sessionId: string | undefined,
-  actor: Record<string, unknown> | undefined = undefined,
-): Record<string, unknown> {
-  return {
-    ...(actor ?? {}),
-    kind: "mcp_provider",
-    ...(sessionId !== undefined && { sessionId }),
-  };
-}
-
-function createResultSummary(output: unknown): string {
-  const outputStr = typeof output === "string" ? output : JSON.stringify(output);
-  const length = outputStr.length;
-  return `success:text:${length}b`;
 }
